\name{dc_arc_pres_abs}
\alias{dc_arc_pres_abs}
\docType{data}
\title{
Stream segment presence absence data for Dry Cr. Idaho 
}
\description{
Stream segment presence absence data for Dry Cr. Idaho (outlet coordinates: 43.71839\eqn{^\circ}N, 116.13747\eqn{^\circ}W).  Arc outcomes determined from STIC (Stream Temperature, Intermittency, and Conductivity) sensors at bounding nodes. 
}
\usage{data("dc_arc_pres_abs")}
\format{
  A data frame with 46187 observations on the following 29 variables.
  \describe{
    \item{\code{datetime}}{a POSIXlt}
    \item{\samp{DC10-->C1}}{a numeric vector}
    \item{\samp{C1-->DC12}}{a numeric vector}
    \item{\samp{DC11-->C1}}{a numeric vector}
    \item{\samp{DC12-->C2}}{a numeric vector}
    \item{\samp{C2-->DC15}}{a numeric vector}
    \item{\samp{DC13-->C2}}{a numeric vector}
    \item{\samp{DC15-->C3}}{a numeric vector}
    \item{\samp{C3-->DC16}}{a numeric vector}
    \item{\samp{DC14-->C3}}{a numeric vector}
    \item{\samp{DC16-->C4}}{a numeric vector}
    \item{\samp{C4-->DC19}}{a numeric vector}
    \item{\samp{DC17-->C5}}{a numeric vector}
    \item{\samp{C5-->C4}}{a numeric vector}
    \item{\samp{DC18-->C5}}{a numeric vector}
    \item{\samp{DC19-->C6}}{a numeric vector}
    \item{\samp{C6-->DC4}}{a numeric vector}
    \item{\samp{DC20-->C6}}{a numeric vector}
    \item{\samp{DC4-->C7}}{a numeric vector}
    \item{\samp{C7-->DC5}}{a numeric vector}
    \item{\samp{DC1-->DC2}}{a numeric vector}
    \item{\samp{DC2-->DC3}}{a numeric vector}
    \item{\samp{DC3-->C7}}{a numeric vector}
    \item{\samp{DC5-->C8}}{a numeric vector}
    \item{\samp{C8-->DC6}}{a numeric vector}
    \item{\samp{DC9-->C9}}{a numeric vector}
    \item{\samp{C9-->DC7}}{a numeric vector}
    \item{\samp{DC8-->C9}}{a numeric vector}
    \item{\samp{DC7-->C8}}{a numeric vector}
  }
}
\source{
Maggie Kraft
}
\keyword{datasets}

