\name{dc_lengths}
\alias{dc_lengths}
\docType{data}
\title{
Lengths of Dry Creek stream (arc) segments
}
\description{
Lengths of stream (arc) segments from Dry Creek Idaho (outlet coordinates: 43.71839\eqn{^\circ}N, 116.13747\eqn{^\circ}W).
}
\usage{data("dc_lengths")}
\format{
  A data frame with 28 observations on the following 2 variables.
  \describe{
    \item{\code{Arcs}}{Arc names, arrows directionally connect nodes.}
    \item{\code{Lengths}}{Stream segment (arc) length in meters.}
  }
}
\source{
Maggie Kraft
}
\keyword{datasets}

