\name{dc_node_pres_abs}
\alias{dc_node_pres_abs}
\docType{data}
\title{
Stream node presence absence data for Dry Cr. Idaho 
}
\description{
Stream node surface water presence absence at Dry Creek ID (outlet coordinates: 43.71839\eqn{^\circ}N, 116.13747\eqn{^\circ}W). Outcomes based on STIC (Stream Temperature, Intermittency, and Conductivity) sensor and piezometer responses, resulting in binary observations for 36 nodes at 15 minutes intervals, over four years.   
}
\usage{data("dc_node_pres_abs")}
\format{
  A data frame with 86958 observations on the following 37 variables.
  \describe{
    \item{\code{datetime}}{a POSIXlt object}
    \item{\code{DC10}}{a numeric vector}
    \item{\code{C1}}{a numeric vector}
    \item{\code{DC11}}{a numeric vector}
    \item{\code{DPZ07}}{a numeric vector}
    \item{\code{DPZ06}}{a numeric vector}
    \item{\code{DC12}}{a numeric vector}
    \item{\code{C2}}{a numeric vector}
    \item{\code{DC13}}{a numeric vector}
    \item{\code{DC15}}{a numeric vector}
    \item{\code{C3}}{a numeric vector}
    \item{\code{DC14}}{a numeric vector}
    \item{\code{DC16}}{a numeric vector}
    \item{\code{DPZ05}}{a numeric vector}
    \item{\code{C4}}{a numeric vector}
    \item{\code{DC17}}{a numeric vector}
    \item{\code{C5}}{a numeric vector}
    \item{\code{DC18}}{a numeric vector}
    \item{\code{DC19}}{a numeric vector}
    \item{\code{C6}}{a numeric vector}
    \item{\code{DC20}}{a numeric vector}
    \item{\code{DC4}}{a numeric vector}
    \item{\code{C7}}{a numeric vector}
    \item{\code{DC1}}{a numeric vector}
    \item{\code{DC2}}{a numeric vector}
    \item{\code{DC3}}{a numeric vector}
    \item{\code{DC5}}{a numeric vector}
    \item{\code{DPZ02}}{a numeric vector}
    \item{\code{C8}}{a numeric vector}
    \item{\code{DPZ04}}{a numeric vector}
    \item{\code{DPZ03}}{a numeric vector}
    \item{\code{DC9}}{a numeric vector}
    \item{\code{C9}}{a numeric vector}
    \item{\code{DC8}}{a numeric vector}
    \item{\code{DC7}}{a numeric vector}
    \item{\code{DC6}}{a numeric vector}
    \item{\code{DSS01}}{a numeric vector}
  }
}
\source{
Maggie Kraft
}
\keyword{datasets}

