\name{gj_coords16}
\alias{gj_coords16}
\docType{data}
\title{
Coordinates of nodes at Gibson Jack Creek, Idaho for a 2016 survey
}
\description{
Latitudes and Longitudes of nodes established at Gibson Jack in 2016.  Datum: WGS 84.
}
\usage{data("gj_coords16")}
\format{
  A data frame with 124 observations on the following 3 variables.
  \describe{
    \item{\code{Object.ID}}{Node name}
    \item{\code{lat}}{Latitude}
    \item{\code{long}}{Longitude}
  }
}
\keyword{datasets}

