\name{gj_lengths}
\alias{gj_lengths}
\docType{data}
\title{
Lengths of Gibson Jack stream (arc) segments
}
\description{
Lengths of Gibson Jack stream (arc) segments
}
\usage{data("gj_lengths")}
\format{
  A data frame with 28 observations on the following 2 variables.
  \describe{
    \item{\code{Arcs}}{Arc names, arrows directionally connect nodes.}
    \item{\code{Lengths}}{Stream segment (arc) length in meters. }
  }
}

\source{
Maggie Kraft
}
\keyword{datasets}

