\name{gj_node_pres_abs}
\alias{gj_node_pres_abs}
\docType{data}
\title{
Stream node presence absence data for Gibson Jack Idaho 
}
\description{
Stream node surface water presence absence data from Gibson Jack, Idaho (outlet coordinates: 42.767180\eqn{^\circ}N, 112.480240\eqn{^\circ}W).  Outcomes based on STIC (Stream Temperature, Intermittency, and Conductivity) sensors, resulting in binary observations for 29 nodes at 15 minutes intervals over three years.   
}
\usage{data("gj_node_pres_abs")}
\format{
  A data frame with 55109 observations on the following 30 variables.
  \describe{
    \item{\code{datetime}}{a POSIXlt}
    \item{\code{GJ16}}{a numeric vector}
    \item{\code{GJ15}}{a numeric vector}
    \item{\code{GJ14}}{a numeric vector}
    \item{\code{C2}}{a numeric vector}
    \item{\code{C3}}{a numeric vector}
    \item{\code{C4}}{a numeric vector}
    \item{\code{GJ11}}{a numeric vector}
    \item{\code{GJ13}}{a numeric vector}
    \item{\code{GJ12}}{a numeric vector}
    \item{\code{GJ20}}{a numeric vector}
    \item{\code{GJ18}}{a numeric vector}
    \item{\code{GJ17}}{a numeric vector}
    \item{\code{C5}}{a numeric vector}
    \item{\code{GJ10}}{a numeric vector}
    \item{\code{GJ9}}{a numeric vector}
    \item{\code{C6}}{a numeric vector}
    \item{\code{GJ23}}{a numeric vector}
    \item{\code{GJ22}}{a numeric vector}
    \item{\code{C1}}{a numeric vector}
    \item{\code{C7}}{a numeric vector}
    \item{\code{GJ24}}{a numeric vector}
    \item{\code{GJ21}}{a numeric vector}
    \item{\code{GJ8}}{a numeric vector}
    \item{\code{GJ7}}{a numeric vector}
    \item{\code{GJ3}}{a numeric vector}
    \item{\code{C8}}{a numeric vector}
    \item{\code{GJ6}}{a numeric vector}
    \item{\code{GJ5}}{a numeric vector}
    \item{\code{GJ4}}{a numeric vector}
  }
}
\source{
Maggie Kraft
}
\keyword{datasets}

