\name{gj_node_pres_abs16}
\alias{gj_node_pres_abs16}
\docType{data}
\title{
Stream node presence absence data for Gibson Jack Cr. Idaho, for a 2016 survey 
}
\description{
Streamflow presence and absence data for each node location collected by manual observation November 6, 2016, May 6, 2017, and August 14, 2017.  Note, a longer dataset 2021-2023, gathered by the AIMS team at fewer points, is available for Gibson Jack under the name \code{gj_node_pres_abs}.  
}
\usage{data("gj_node_pres_abs16")}
\format{
  A data frame with 3 observations on the following 125 variables.
  \describe{
    \item{\code{Date}}{a character vector}
    \item{\code{GJ_ST1_0600}}{a numeric vector}
    \item{\code{GJ_ST1_0400}}{a numeric vector}
    \item{\code{GJ_ST1_0200}}{a numeric vector}
    \item{\code{GJ_ST1_0000}}{a numeric vector}
    \item{\code{GJ_SF_2800}}{a numeric vector}
    \item{\code{GJ_SF_2600}}{a numeric vector}
    \item{\code{GJ_SF_2400}}{a numeric vector}
    \item{\code{GJ_SF_2200}}{a numeric vector}
    \item{\code{GJ_SF_2000}}{a numeric vector}
    \item{\code{GJ_SF_1800}}{a numeric vector}
    \item{\code{GJ_SF_1600}}{a numeric vector}
    \item{\code{GJ_SF_1400}}{a numeric vector}
    \item{\code{GJ_SF_1200}}{a numeric vector}
    \item{\code{GJ_SF_1000}}{a numeric vector}
    \item{\code{GJ_SF_0800}}{a numeric vector}
    \item{\code{GJ_SF_0600}}{a numeric vector}
    \item{\code{GJ_SF_0400}}{a numeric vector}
    \item{\code{GJ_SF_0200}}{a numeric vector}
    \item{\code{GJ_SF_0000}}{a numeric vector}
    \item{\code{GJ_NT1_WF_FH}}{a numeric vector}
    \item{\code{GJ_NT1_WF_000}}{a numeric vector}
    \item{\code{GJ_NT1_0800}}{a numeric vector}
    \item{\code{GJ_NT1_0600}}{a numeric vector}
    \item{\code{GJ_NT1_0400}}{a numeric vector}
    \item{\code{GJ_NT1_0200}}{a numeric vector}
    \item{\code{GJ_NT1_0000}}{a numeric vector}
    \item{\code{GJ_NT2_1600}}{a numeric vector}
    \item{\code{GJ_NT2_1400}}{a numeric vector}
    \item{\code{GJ_NT2_1200}}{a numeric vector}
    \item{\code{GJ_NT2_1000}}{a numeric vector}
    \item{\code{GJ_NT2_0800}}{a numeric vector}
    \item{\code{GJ_NT2_0600}}{a numeric vector}
    \item{\code{GJ_NT2_0400}}{a numeric vector}
    \item{\code{GJ_NT2_0200}}{a numeric vector}
    \item{\code{GJ_NT2_0000}}{a numeric vector}
    \item{\code{GJ_NT3_0200}}{a numeric vector}
    \item{\code{GJ_NT3_0000}}{a numeric vector}
    \item{\code{GJ_NT4_0600}}{a numeric vector}
    \item{\code{GJ_NT4_0400}}{a numeric vector}
    \item{\code{GJ_NT4_0200}}{a numeric vector}
    \item{\code{GJ_NT4_0000}}{a numeric vector}
    \item{\code{GJ_NF_3800}}{a numeric vector}
    \item{\code{GJ_NF_3750}}{a numeric vector}
    \item{\code{GJ_NF_3600}}{a numeric vector}
    \item{\code{GJ_NF_3400}}{a numeric vector}
    \item{\code{GJ_NF_3200}}{a numeric vector}
    \item{\code{GJ_NF_3000_CU}}{a numeric vector}
    \item{\code{GJ_NF_3000_CD}}{a numeric vector}
    \item{\code{GJ_NF_2800_CU}}{a numeric vector}
    \item{\code{GJ_NF_2800_CD}}{a numeric vector}
    \item{\code{GJ_NF_2600}}{a numeric vector}
    \item{\code{GJ_NF_2400_CU}}{a numeric vector}
    \item{\code{GJ_NF_2400_CD}}{a numeric vector}
    \item{\code{GJ_NF_2200}}{a numeric vector}
    \item{\code{GJ_NF_2000}}{a numeric vector}
    \item{\code{GJ_NF_1800}}{a numeric vector}
    \item{\code{GJ_NF_1600}}{a numeric vector}
    \item{\code{GJ_NF_1400}}{a numeric vector}
    \item{\code{GJ_NF_1200}}{a numeric vector}
    \item{\code{GJ_NF_1060_CU}}{a numeric vector}
    \item{\code{GJ_NF_1060_CD}}{a numeric vector}
    \item{\code{GJ_NF_1000}}{a numeric vector}
    \item{\code{GJ_NF_0800}}{a numeric vector}
    \item{\code{GJ_NF_0600}}{a numeric vector}
    \item{\code{GJ_NF_0400}}{a numeric vector}
    \item{\code{GJ_NF_0200}}{a numeric vector}
    \item{\code{GJ_NF_0000}}{a numeric vector}
    \item{\code{GJ_MT2_0900}}{a numeric vector}
    \item{\code{GJ_MT2_0800}}{a numeric vector}
    \item{\code{GJ_MT2_0600}}{a numeric vector}
    \item{\code{GJ_MT2_0400}}{a numeric vector}
    \item{\code{GJ_MT2_0200}}{a numeric vector}
    \item{\code{GJ_MT2_0000}}{a numeric vector}
    \item{\code{GJ_MT1_1650}}{a numeric vector}
    \item{\code{GJ_MT1_1600}}{a numeric vector}
    \item{\code{GJ_MT1_1550}}{a numeric vector}
    \item{\code{GJ_MT1_1500}}{a numeric vector}
    \item{\code{GJ_MT1_1450}}{a numeric vector}
    \item{\code{GJ_MT1_1400}}{a numeric vector}
    \item{\code{GJ_MT1_1350}}{a numeric vector}
    \item{\code{GJ_MT1_1300}}{a numeric vector}
    \item{\code{GJ_MT1_1250}}{a numeric vector}
    \item{\code{GJ_MT1_1200}}{a numeric vector}
    \item{\code{GJ_MT1_1150}}{a numeric vector}
    \item{\code{GJ_MT1_1100}}{a numeric vector}
    \item{\code{GJ_MT1_1050}}{a numeric vector}
    \item{\code{GJ_MT1_1000}}{a numeric vector}
    \item{\code{GJ_MT1_0950}}{a numeric vector}
    \item{\code{GJ_MT1_0900}}{a numeric vector}
    \item{\code{GJ_MT1_0850}}{a numeric vector}
    \item{\code{GJ_MT1_0800}}{a numeric vector}
    \item{\code{GJ_MT1_0750}}{a numeric vector}
    \item{\code{GJ_MT1_0700}}{a numeric vector}
    \item{\code{GJ_MT1_0650}}{a numeric vector}
    \item{\code{GJ_MT1_0600}}{a numeric vector}
    \item{\code{GJ_MT1_0550}}{a numeric vector}
    \item{\code{GJ_MT1_0500}}{a numeric vector}
    \item{\code{GJ_MT1_0450}}{a numeric vector}
    \item{\code{GJ_MT1_0400}}{a numeric vector}
    \item{\code{GJ_MT1_0350}}{a numeric vector}
    \item{\code{GJ_MT1_0300}}{a numeric vector}
    \item{\code{GJ_MT1_0250}}{a numeric vector}
    \item{\code{GJ_MT1_0200}}{a numeric vector}
    \item{\code{GJ_MT1_0150}}{a numeric vector}
    \item{\code{GJ_MT1_0100}}{a numeric vector}
    \item{\code{GJ_MT1_0050}}{a numeric vector}
    \item{\code{GJ_MT1_0000}}{a numeric vector}
    \item{\code{GJ_MS_3000}}{a numeric vector}
    \item{\code{GJ_MS_2800}}{a numeric vector}
    \item{\code{GJ_MS_2600}}{a numeric vector}
    \item{\code{GJ_MS_2400}}{a numeric vector}
    \item{\code{GJ_MS_2200}}{a numeric vector}
    \item{\code{GJ_MS_2000}}{a numeric vector}
    \item{\code{GJ_MS_1800_CU}}{a numeric vector}
    \item{\code{GJ_MS_1800_CD}}{a numeric vector}
    \item{\code{GJ_MS_1600}}{a numeric vector}
    \item{\code{GJ_MS_1400}}{a numeric vector}
    \item{\code{GJ_MS_1200}}{a numeric vector}
    \item{\code{GJ_MS_1000}}{a numeric vector}
    \item{\code{GJ_MS_0800}}{a numeric vector}
    \item{\code{GJ_MS_0600}}{a numeric vector}
    \item{\code{GJ_MS_0400}}{a numeric vector}
    \item{\code{GJ_MS_0200}}{a numeric vector}
    \item{\code{GJ_MS_0000}}{a numeric vector}
  }
}
\keyword{datasets}

