\name{jd_lengths}
\alias{jd_lengths}
\docType{data}
\title{
Lengths of Johnson Draw stream (arc) segments
}
\description{
Lengths of Johnson Draw stream (arc) segments
}
\usage{data("jd_lengths")}
\format{
  A data frame with 45 observations on the following 2 variables.
  \describe{
    \item{\code{Arcs}}{Arc names, arrows directionally connect nodes.}
    \item{\code{Lengths}}{Lengths in meters}
  }
}
\source{
Arya Legg, Maggie Kraft
}
\keyword{datasets}

