\name{jd_node_pres_abs}
\alias{jd_node_pres_abs}
\docType{data}
\title{
Stream node presence absence data for Johnson Draw Idaho 
}
\description{
Stream node surface water presence absence data from 2022-2023 for the Johnson Draw watershed in southwest Idaho (outlet coordinates: 43.12256\eqn{^\circ}N, 116.77630\eqn{^\circ}W).  Outcomes are based on STIC sensors and piezometers, resulting in binary observations for 35 nodes (28 STICs and 7 piezometers) at 15 minutes intervals over three years. JD21, JD22, JD23, JD25, JD26, JD27, JD28 were added in 2023 for the AIMS experiment.  
}
\usage{data("jd_node_pres_abs")}
\format{
  A data frame with 51653 observations on the following 36 variables.
  \describe{
    \item{\code{datetime}}{a POSIXlt}
    \item{\code{JD5}}{a numeric vector}
    \item{\code{JD6}}{a numeric vector}
    \item{\code{JD7}}{a numeric vector}
    \item{\code{C1}}{a numeric vector}
    \item{\code{JD10}}{a numeric vector}
    \item{\code{JD9}}{a numeric vector}
    \item{\code{JD8}}{a numeric vector}
    \item{\code{JD11}}{a numeric vector}
    \item{\code{JD12}}{a numeric vector}
    \item{\code{JD13}}{a numeric vector}
    \item{\code{C2}}{a numeric vector}
    \item{\code{JD16}}{a numeric vector}
    \item{\code{JD17}}{a numeric vector}
    \item{\code{JD18}}{a numeric vector}
    \item{\code{JD19}}{a numeric vector}
    \item{\code{JD20}}{a numeric vector}
    \item{\code{JD4}}{a numeric vector}
    \item{\code{JD3}}{a numeric vector}
    \item{\code{JD2}}{a numeric vector}
    \item{\code{JD1}}{a numeric vector}
    \item{\code{JD15}}{a numeric vector}
    \item{\code{JD14}}{a numeric vector}
    \item{\code{JSS01}}{a numeric vector}
    \item{\code{JPZ02}}{a numeric vector}
    \item{\code{JPZ03}}{a numeric vector}
    \item{\code{JPZ04}}{a numeric vector}
    \item{\code{JPZ05}}{a numeric vector}
    \item{\code{JPZ06}}{a numeric vector}
    \item{\code{JPZ07}}{a numeric vector}
    \item{\code{JD21}}{a numeric vector}
    \item{\code{JD22}}{a numeric vector}
    \item{\code{JD23}}{a numeric vector}
    \item{\code{JD25}}{a numeric vector}
    \item{\code{JD26}}{a numeric vector}
    \item{\code{JD27}}{a numeric vector}
    \item{\code{JD28}}{a numeric vector}
    
    
    
  }
}

\source{
Maggie Kraft
}
\keyword{datasets}

