\name{kon_lengths}
\alias{kon_lengths}
\docType{data}
\title{
Lengths of Murphy Cr. stream (arc) segments
}
\description{
Lengths of Murphy Cr. stream (arc) segments
}
\usage{data("kon_lengths")}
\format{
  A data frame with 45 observations on the following 2 variables.
  \describe{
    \item{\code{Arcs}}{Arc names, arrows directionally connect nodes.}
    \item{\code{Lengths}}{Lengths in meters}
  }
}
\source{
Rob Ramos
}
\keyword{datasets}

