\name{kon_coords}
\alias{kon_coords}
\docType{data}
\title{
Coordinates of nodes in the Konza Praire dataset
}
\description{
Coordinates (in Lat/Long) of nodes established at the Konza Prairie stream network.
}
\usage{data("kon_coords")}
\format{
  A data frame with 46 observations on the following 3 variables.
  \describe{
    \item{\code{Object.ID}}{Node name}
    \item{\code{lat}}{Latitude}
    \item{\code{long}}{Longitude}
  }
}
\keyword{datasets}

