\name{mur_arc_pres_abs}
\alias{mur_arc_pres_abs}
\docType{data}
\title{
Stream segment presence absence data for Murphy Cr. Idaho 
}
\description{
Simulated multivariate Benroulli outcomes for 27 stream segments, based on their observed marginal probabilities for steam presence and covariance structures. "M"-labelling for nodes indicates "meters above outlet".  
}
\usage{data("mur_arc_pres_abs")}
\format{
  A data frame with 1000 observations on the following 27 variables.
  \describe{
    \item{\samp{IN_N-->M1984}}{a numeric vector}
    \item{\samp{M1984-->M1909}}{a numeric vector}
    \item{\samp{M1909-->M1799}}{a numeric vector}
    \item{\samp{IN_S-->M1993}}{a numeric vector}
    \item{\samp{M1993-->M1951}}{a numeric vector}
    \item{\samp{M1951-->M1909}}{a numeric vector}
    \item{\samp{M1799-->M1719}}{a numeric vector}
    \item{\samp{M1719-->M1653}}{a numeric vector}
    \item{\samp{M1653-->M1572}}{a numeric vector}
    \item{\samp{M1572-->M1452}}{a numeric vector}
    \item{\samp{M1452-->M1377}}{a numeric vector}
    \item{\samp{M1377-->M1254}}{a numeric vector}
    \item{\samp{M1254-->M1166}}{a numeric vector}
    \item{\samp{M1166-->M1121}}{a numeric vector}
    \item{\samp{M1121-->M1036}}{a numeric vector}
    \item{\samp{M1036-->M918}}{a numeric vector}
    \item{\samp{M918-->M823}}{a numeric vector}
    \item{\samp{M823-->M759}}{a numeric vector}
    \item{\samp{M759-->M716}}{a numeric vector}
    \item{\samp{M716-->M624}}{a numeric vector}
    \item{\samp{M624-->M523}}{a numeric vector}
    \item{\samp{M523-->M454}}{a numeric vector}
    \item{\samp{M454-->M380}}{a numeric vector}
    \item{\samp{M380-->M233}}{a numeric vector}
    \item{\samp{M233-->M153}}{a numeric vector}
    \item{\samp{M153-->M91}}{a numeric vector}
    \item{\samp{M91-->OUT}}{a numeric vector}
  }
}
\keyword{datasets}

