\name{mur_coords}
\alias{mur_coords}
\docType{data}
\title{
Coordinates of nodes at Murphy Ck. Idaho 
}
\description{
Coordinates in 12T UTMs of nodes established at Murphy Cr. Idaho.
}
\usage{data("mur_coords")}
\format{
  A data frame with 28 observations on the following 3 variables.
  \describe{
    \item{\code{Object.ID}}{Node name}
    \item{\code{x}}{UTM Easting}
    \item{\code{y}}{UTM Northing}
  }
}
\keyword{datasets}

