\name{mur_lengths}
\alias{mur_lengths}
\docType{data}
\title{
Lengths of Murphy Cr. stream (arc) segments
}
\description{
Lengths of Murphy Cr. stream (arc) segments
}
\usage{data("mur_lengths")}
\format{
  A data frame with 27 observations on the following 2 variables.
  \describe{
    \item{\code{Arcs}}{Arc names, arrows directionally connect nodes.}
    \item{\code{Lengths}}{Stream segment (arc) length in meters. }
  }
}
\source{
Warix, S. R., Godsey, S. E., Lohse, K. A., & Hale, R. L. (2021). Influence of groundwater and topography on stream drying in semi-arid headwater streams. Hydrological Processes, 35(5), e14185.
}
\keyword{datasets}

