\name{mur_node_pres_abs}
\alias{mur_node_pres_abs}
\docType{data}
\title{
Stream node presence absence data for Murphy Cr. Idaho 
}
\description{
A subset of stream node presence absence data from Warix et al. (2019) resulting in binary observations for 28 nodes at 2.5 hr intervals.  
}
\usage{data("mur_node_pres_abs")}
\format{
  A data frame with 1163 observations on the following 29 variables.
  \describe{
    \item{\code{Datetime}}{a character vector}
    \item{\code{IN_N}}{a numeric vector}
    \item{\code{M1984}}{a numeric vector}
    \item{\code{M1909}}{a numeric vector}
    \item{\code{IN_S}}{a numeric vector}
    \item{\code{M1993}}{a numeric vector}
    \item{\code{M1951}}{a numeric vector}
    \item{\code{M1799}}{a numeric vector}
    \item{\code{M1719}}{a numeric vector}
    \item{\code{M1653}}{a numeric vector}
    \item{\code{M1572}}{a numeric vector}
    \item{\code{M1452}}{a numeric vector}
    \item{\code{M1377}}{a numeric vector}
    \item{\code{M1254}}{a numeric vector}
    \item{\code{M1166}}{a numeric vector}
    \item{\code{M1121}}{a numeric vector}
    \item{\code{M1036}}{a numeric vector}
    \item{\code{M918}}{a numeric vector}
    \item{\code{M823}}{a numeric vector}
    \item{\code{M759}}{a numeric vector}
    \item{\code{M716}}{a numeric vector}
    \item{\code{M624}}{a numeric vector}
    \item{\code{M523}}{a numeric vector}
    \item{\code{M454}}{a numeric vector}
    \item{\code{M380}}{a numeric vector}
    \item{\code{M233}}{a numeric vector}
    \item{\code{M153}}{a numeric vector}
    \item{\code{M91}}{a numeric vector}
    \item{\code{OUT}}{a numeric vector}
  }
}
\references{
Warix, S. R., Godsey, S. E., Lohse, K. A., & Hale, R. L. (2021). Influence of groundwater and topography on stream drying in semi-arid headwater streams. \emph{Hydrological Processes, 35}(5), e14185.
}
\keyword{datasets}

