\name{ta_coords}
\alias{ta_coords}
\docType{data}
\title{
Coordinates of nodes at Talledega Creek, Alabama
}
\description{
Latitudes and Longitudes of nodes established at Talledega Creek in 2022.  Datum: WGS 84.
}
\usage{data("ta_coords")}
\format{
  A data frame with 56 observations on the following 3 variables.
  \describe{
    \item{\code{Object.ID}}{Node name}
    \item{\code{lat}}{Latitude}
    \item{\code{long}}{Longitude}
  }
}
\keyword{datasets}

