\name{gj_lengths}
\alias{gj_lengths}
\docType{data}
\title{
Lengths of Gibson Jack stream (arc) segments
}
\description{
Lengths of stream (arc) segments from the Gibson Jack watershed in southeast Idaho (outlet coordinates: 42.767180\eqn{^\circ}N, 112.480240\eqn{^\circ}W).
}
\usage{data("gj_lengths")}
\format{
  A data frame with 28 observations on the following 2 variables.
  \describe{
    \item{\code{Arcs}}{Arc names, arrows directionally connect nodes.}
    \item{\code{Lengths}}{Stream segment (arc) length in meters. }
  }
}

\source{
Maggie Kraft
}
\keyword{datasets}

