\name{jd_node_pres_abs}
\alias{jd_node_pres_abs}
\docType{data}
\title{
Stream node presence absence data for Murphy Cr. Idaho 
}
\description{
Stream node surface water presence absence data resulting in binary observations for 23 nodes at 15 minutes intervals.  
}
\usage{data("jd_node_pres_abs")}
\format{
  A data frame with 50322 observations on the following 23 variables.
  \describe{
    \item{\code{datetime}}{a character vector}
    \item{\code{JD5}}{a numeric vector}
    \item{\code{JD6}}{a numeric vector}
    \item{\code{JD7}}{a numeric vector}
    \item{\code{C1}}{a numeric vector}
    \item{\code{JD10}}{a numeric vector}
    \item{\code{JD9}}{a numeric vector}
    \item{\code{JD8}}{a numeric vector}
    \item{\code{JD11}}{a numeric vector}
    \item{\code{JD12}}{a numeric vector}
    \item{\code{JD13}}{a numeric vector}
    \item{\code{C2}}{a numeric vector}
    \item{\code{JD16}}{a numeric vector}
    \item{\code{JD17}}{a numeric vector}
    \item{\code{JD18}}{a numeric vector}
    \item{\code{JD19}}{a numeric vector}
    \item{\code{JD20}}{a numeric vector}
    \item{\code{JD4}}{a numeric vector}
    \item{\code{JD3}}{a numeric vector}
    \item{\code{JD2}}{a numeric vector}
    \item{\code{JD1}}{a numeric vector}
    \item{\code{JD15}}{a numeric vector}
    \item{\code{JD14}}{a numeric vector}
  }
}
\source{
Maggie Kraft
}
\keyword{datasets}

