\name{mur_coords}
\alias{mur_coords}
\docType{data}
\title{
Coordinates of nodes at Murphy Ck. Idaho 
}
\description{
UTM coordinates (Zone 11T) and Latitudes and Longitudes of nodes established at Murphy Cr. Idaho.  Datum: WGS 84.
}
\usage{data("mur_coords")}
\format{
  A data frame with 28 observations on the following 5 variables.
  \describe{
    \item{\code{Object.ID}}{Node name}
    \item{\code{E}}{UTM Easting}
    \item{\code{N}}{UTM Northing}
    \item{\code{lat}}{Latitude}
    \item{\code{long}}{Longitude}
  }
}
\keyword{datasets}

